fix: u8 overflow in detect_delete_range for complement sets
When complement of a small set (e.g., -cd '0-9') produces chars spanning
0-255, the expression (hi - lo + 1) overflows u8 in debug mode. Fix by
casting to usize before the +1 arithmetic.

+++ src/tr/core.rs
@@ -2086,7 +2086,8 @@ fn detect_delete_range(chars: &[u8]) -> Option<(u8, u8)> {
         }
     }
     // Check that the range size matches the number of chars (no gaps)
-    if (hi - lo + 1) as usize == chars.len() {
+    // Cast to usize before +1 to avoid u8 overflow when hi=255, lo=0 (range=256)
+    if (hi as usize - lo as usize + 1) == chars.len() {
         Some((lo, hi))
     } else {
         None
